fixed clean up containers script

+++ clean_up_containers.sh
@@ -1,1 +1,4 @@
-docker rmi -f $(docker images -f "dangling=true" -q)
+if docker images -f "dangling=true" -q | wc -l != 0
+  then
+	docker rmi -f $(docker images -f "dangling=true" -q)
+fi
